A wrapper class for DrupalContext to allow Composition instead of Extension
This is a hack, but it is done with a pure spirit! (puke)
DrupalContext's (and really, Mink's) locatePath method is written with the
assumption of class hierarchy instead of object composition (it is protected).
This does not play well with our dynamic subcontext loading.

The motivation here is to allow the Developer Experience to be consistent
with the former structure (i.e. they can still use $this->locatePath() istead of
$this->getMainContext()->getSubcontextByClassName('\Drupal\DrupalExtension\Context\DrupalContext')->locatePath()